tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.get.order.alpha.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.get.order.alpha</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.get.order.alpha</h1>
<p class="desc">getImageData() returns A in the fourth component</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("getImageData() returns A in the fourth component");
_addTest(function(canvas, ctx) {

ctx.fillStyle = 'rgba(0, 0, 0, 0.5)';
ctx.fillRect(0, 0, 100, 50);
var imgdata = ctx.getImageData(0, 0, 10, 10);
_assert(imgdata.data[3] < 200, "imgdata.data[\""+(3)+"\"] < 200");
_assert(imgdata.data[3] > 100, "imgdata.data[\""+(3)+"\"] > 100");


});
</script>
